requirements.txt
llama-cpp-python
transformers
whisper
vosk
